login ver with check dorm

+++ Java/DormitoryNotebooksJava/src/main/java/kaczuch/master_thesis/repositories/DormRepository.java
@@ -12,7 +12,7 @@ import java.util.Optional;
 public interface DormRepository extends JpaRepository<Dorm, Integer> {
 
     Optional<Dorm> findByName(String name);
-    Optional<Dorm> findById(Integer id);
+    Optional<Dorm> findById(Long id);
 
     List<Dorm> findByUsers_Id(Long userId);
 }
